Massive refactors/changes to everything
Animus network now is bound to the player not the world.
AnimusHelper is now the defacto way to control animus for anything.
Moved some helpers and libs into the api.
The Blueprint now cares about items on the ground, not items in the
blueprint. It is no longer an IInventory.